DRY fix

+++ Exercises/Franks_Sticky_Calculator.py
@@ -1,10 +1,10 @@
 def sticky_calc(operation, val1, val2):
-    stick = str(round(val1)) + str(round(val2))
+    stick = int(str(round(val1)) + str(round(val2)))
     if operation == '+':
-        return round(int(stick) + val2)
+        return round(stick + val2)
     elif operation == '-':
-        return round(int(stick) - val2)
+        return round(stick - val2)
     elif operation == '*':
-        return round(int(stick) * val2)
+        return round(stick * val2)
     elif operation == '/':
-        return round(int(stick) / val2)
+        return round(stick / val2)
